Refactor materials
