+++ CgfConverter/Materials/SubMaterials.cs
@@ -1,0 +1,11 @@
+﻿using System.Collections.Generic;
+using System.Xml.Serialization;
+
+namespace CgfConverter.Materials;
+
+[XmlRoot(ElementName = "SubMaterials")]
+public class SubMaterials
+{
+    [XmlElement(ElementName = "Material")]
+    public List<Material> Material { get; set; } = new();
+}
