+++ CgfConverter/Materials/Textures.cs
@@ -1,0 +1,11 @@
+﻿using System.Collections.Generic;
+using System.Xml.Serialization;
+
+namespace CgfConverter.Materials;
+
+[XmlRoot(ElementName = "Textures")]
+public record Textures
+{
+    [XmlElement(ElementName = "Texture")]
+    readonly List<Texture> Texture = new ();
+}
